+++ 3w_ticketing/src/main/java/com/hhplus/concert_ticketing/domain/user/UserRepository.java
@@ -3,17 +3,17 @@ package com.hhplus.concert_ticketing.domain.user;
 import java.util.Optional;
 
 public interface UserRepository {
-    Optional<Double> findBalanceByUserId(Long userId);
+    Optional<Integer> findBalanceByUserId(Long userId);
 
     Optional<UserEntity> findByIdForUpdate(Long userId);
 
-    int chargePoint(Long userId, double amount);
+    int chargePoint(Long userId, int amount);
 
     Optional<UserEntity> findById(Long userId);
 
     UserEntity save(UserEntity user);
 
-    int usePoint(Long userId, Double balance, int version);
+    int usePoint(Long userId, int balance, int version);
 
     UserEntity getUserInfo(Long userId);
 }
